requirements.txt
streamlit==1.31.0
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==1.4.50
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pandas==2.1.3
numpy==1.26.2
plotly==5.18.0
scikit-learn==1.3.2
pytest==7.4.3
black==23.11.0